refactor: adiciona tipos para o pacote cookie e corrige erros de logout

+++ src/modules/authentication/application/useCase/requestRefreshToken.useCase.ts
@@ -1,5 +1,3 @@
-import bcrypt from 'bcrypt';
-import { sign, type Secret } from 'jsonwebtoken';
 import { AppError } from '../../../../core/application/AppError';
 import { type IUseCase } from '../../../../core/application/useCase/IUseCase';
 import DataExpiracao from '../../../../core/domain/valueObjects/DataExpiracao';
@@ -13,6 +11,8 @@ import { Token } from '../../domain/entities/Token';
 import type ITokenRepository from '../../domain/repositories/ITokenRepository';
 import type Usuario from '../../../usuario/domain/entities/Usuario';
 import { RequestRefreshTokenErrors } from './requestRefreshTokenErrors';
+import { getUnixTime, isAfter } from 'date-fns';
+import TokenVO from '../../../../core/domain/valueObjects/Token';
 
 type Response = Either<
   AppError.UnexpectedError,
@@ -40,31 +40,22 @@ export class RequestRefreshTokenUseCase
       if (tokenData.isLeft()) {
         return left(new RequestRefreshTokenErrors.TokenInvalid());
       }
+      const tokenDataExpired = isAfter(
+        new Date(),
+        getUnixTime(parseInt(tokenData.value.getValue().dataExpiracao)),
+      );
 
-      await this.tokenRepository.delete(tokenData.value.getValue());
+      if (tokenDataExpired) {
+        await this.tokenRepository.delete(tokenData.value.getValue());
+      }
 
-      const { JWT_SECRET } = process.env;
-
-      const tokenJWT = sign(
-        {
+      const token = Token.create({
+        token: await TokenVO.setValue({
           id: tokenData.value.getValue().id.toString(),
           nome: input.usuario.nome,
           email: input.usuario.email,
-          tenantId: tokenData.value.getValue().tenantId,
-        },
-        JWT_SECRET as Secret,
-        {
-          subject: tokenData.value.getValue().id.toString(),
-          expiresIn: '1d',
-        },
-      );
-
-      const salt = await bcrypt.genSalt(12);
-
-      const tokenHash = await bcrypt.hash(tokenJWT, salt);
-
-      const token = Token.create({
-        token: tokenHash,
+          tenantId: input.usuario.tenantId,
+        }),
         usuarioId: input.usuario.id.toString(),
         tenantId: input.usuario.tenantId,
         dataExpiracao: DataExpiracao.setValue(),
@@ -78,7 +69,7 @@ export class RequestRefreshTokenUseCase
         Result.ok<{
           token: string;
         }>({
-          token: tokenJWT,
+          token: token.token,
         }),
       );
     } catch (error) {
